+++ test/AtA_AAt.c
@@ -6,12 +6,12 @@
 #define M 3
 
 int main() {
-    float A[M*N] = {3, 1, 1, 2, 0, 2, 0, 1, 2, 1, 3, 2};
-    float exp_AtA[N*N] = {13, 5, 9, 10, 5, 6, 4, 6, 9, 4, 10, 8, 10, 6, 8, 9};
-    float exp_AAt[M*M] = {15, 4, 14, 4, 5, 4, 14, 4, 18};
+    double A[M*N] = {3, 1, 1, 2, 0, 2, 0, 1, 2, 1, 3, 2};
+    double exp_AtA[N*N] = {13, 5, 9, 10, 5, 6, 4, 6, 9, 4, 10, 8, 10, 6, 8, 9};
+    double exp_AAt[M*M] = {15, 4, 14, 4, 5, 4, 14, 4, 18};
 
-    float *AtA = get_AtA(A, M, N);
-    float *AAt = get_AAt(A, M, N);
+    double *AtA = get_AtA(A, M, N);
+    double *AAt = get_AAt(A, M, N);
 
     // Test AtA
     for (int i=0 ; i<N ; i++)
